Swapped out search system + bug fixes
