+++ packages/core/src/actions/types/search/tracedRecursiveSearch/simpleSearch/fuzzySearchMethod/fuzzySearchMethod.ts
@@ -9,6 +9,7 @@ export const fuzzySearchMethod: ISimpleSearchMethod = {
     name: "Fuzzy",
     ID: "CoreFuzzySearchMethod",
     view: createStandardMenuItem({name: "Fuzzy finder"}),
-    rate: (data, query) => FuzzyRater.getRater(query).rate(data),
-    highlight: (text, query) => FuzzyRater.getRater(query).highlight(text),
+    rate: (data, search, query) => FuzzyRater.getRater(query, search).rate(data),
+    highlight: (text, search, query) =>
+        FuzzyRater.getRater(query, search).highlight(text),
 };
